MXS-2928: Serve files via REST API
The files in the gui subdirectory of the MaxScale sharedir can now be
served to clients like a normal webserver. As the GUI itself performs
authentication, these files do not require it. This is safe as there's no
data in the GUI files themselves and the actual data is requested via the
normal REST API endpoints.

The files are verified to be in the sharedir so that relative URLs do not
access files outside of the GUI files. Although most browsers and clients
correctly interpret relative URLs, it must still be done to prevent
malicious use.